House_cut_apart_code/code/requirements.txt
absl-py
addict
altair
attrs
blinker
cachetools
captcha
certifi
charset-normalizer
click
colorama
coloredlogs
contourpy
cycler
filelock
flatbuffers
fonttools
gitdb
GitPython
google-auth
google-auth-oauthlib
grpcio
humanfriendly
idna
IMcore
importlib-metadata
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
networkx
numpy
oauthlib
onnxruntime
opencv-python
packaging
pandas
Pillow
protobuf
psutil
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pycocotools
pydeck
Pygments
pyparsing
pyreadline3
PySide6
PySide6-Addons
PySide6-Essentials
python-dateutil
pytz
PyYAML
QtFusion
referencing
requests
requests-oauthlib
rich
rsa
scipy
seaborn
shiboken6
six
smmap
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
thop
toml
toolz
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado
tqdm
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
Werkzeug
zipp
timm
einops
efficientnet_pytorch
importlib
dill
